\name{acme}
\alias{acme}
\title{
Monthly Excess Returns
}
\section{SUMMARY}{
The \code{acme} data frame has 60 rows and 3 columns.


The excess return for the Acme Cleveland Corporation are recorded along with
those for all stocks listed on the New York and American Stock Exchanges were
recorded over a five year period.  These excess returns are relative to the
return on a risk-less investment such a U.S. Treasury bills.  
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{month}{
A character string representing the month of the observation.
}
\item{market}{
The excess return of the market as a whole.
}
\item{acme}{
The excess return for the Acme Cleveland Corporation.
}}
\section{SOURCE}{
The data were obtained from


Simonoff, J.S. and Tsai, C.-L. (1994) Use of modified profile likelihood for 
improved tests of constancy of variance in regression. 
\emph{Applied Statistics}, \bold{43}, 353-370.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) \emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
