\name{aircondit}
\alias{aircondit}
\alias{aircondit7}
\title{
Failures of Air-conditioning Equipment
}
\section{SUMMARY}{
Proschan (1963) reported on the times between failures of the air-conditioning
equipment in 10 Boeing 720 aircraft. The \code{aircondit} data frame contains 
the intervals for the ninth aircraft while \code{aircondit7} contains those for the
seventh aircraft. 


Both data frames have just one column.  Note that the data have been sorted 
into increasing order.
}
\section{DATA DESCRIPTION}{
The data frames contain the following column:


}
\arguments{
\item{hours}{
The time interval in hours between successive failures of the air-conditioning
equipment
}}
\section{SOURCE}{
The data were taken from


Cox, D.R. and Snell, E.J. (1981) 
\emph{Applied Statistics: Principles and Examples}. Chapman and Hall.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Proschan, F. (1963) Theoretical explanation of observed decreasing failure 
rate. \emph{Technometrics}, \bold{5}, 375-383.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
