\name{bigcity}
\alias{bigcity}
\alias{city}
\title{
Population of U.S. Cities
}
\description{
The \code{bigcity} data frame has 49 rows and 2 columns. 


The \code{city} data frame has 10 rows and 2 columns.  


The measurements are the
population (in 1000's) of 49 U.S. cities in 1920 and 1930.  The 49 cities are 
a random sample taken from the 196 largest cities in 1920.  The \code{city} data 
frame consists of the first 10 observations in \code{bigcity}. 
}
\usage{
bigcity
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{u}}{
The 1920 population.
}
\item{\code{x}}{
The 1930 population.
}}}
\source{
The data were obtained from


Cochran, W.G. (1977) \emph{Sampling Techniques}. Third edition. John Wiley
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) \emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
