\name{brambles}
\alias{brambles}
\title{
Spatial Location of Bramble Canes
}
\section{SUMMARY}{
The \code{brambles} data frame has 823 rows and 3 columns.


The location of living bramble canes in a 9m square plot was recorded.  
We take 9m to be the unit of distance so that the plot can be thought of
as a unit square. The bramble canes were also classified by their age.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{x}{
The x coordinate of the position of the cane in the plot.
}
\item{y}{
The y coordinate of the position of the cane in the plot.
}
\item{age}{
The age classification of the canes; \code{0} indicates a newly emerged cane,
\code{1} indicates a one year old cane and \code{2} indicates a two year old cane.
}}
\section{SOURCE}{
The data were obtained from


Diggle, P.J. (1983) \emph{Statistical Analysis of Spatial Point Patterns}.
Academic Press.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
