\name{calcium}
\alias{calcium}
\title{
Calcium Uptake Data
}
\description{
The \code{calcium} data frame has 27 rows and 2 columns.


Howard Grimes from the Botany Department, North Carolina State University,
conducted an experiment for biochemical analysis of intracellular storage
and transport of calcium across plasma membrane.  Cells were
suspended in a solution of radioactive calcium for a certain length of time and
then the amount of radioactive calcium that was absorbed by the cells was
measured.  The experiment was repeated independently with 9 different
times of suspension each replicated 3 times.  
}
\usage{
calcium
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{time}}{
The time (in minutes) that the cells were suspended in the solution.
}
\item{\code{cal}}{
The amount of calcium uptake (nmoles/mg).
}}}
\source{
The data were obtained from


Rawlings, J.O. (1988) \emph{Applied Regression Analysis}. 
Wadsworth and Brooks/Cole Statistics/Probability Series.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
