\name{cats}
\alias{cats}
\title{
Weight Data for Domestic Cats
}
\section{SUMMARY}{
The \code{cats} data frame has 144 rows and 3 columns.


The \code{catsM} data frame has 97 rows and 3 columns.


144 adult (over 2kg in weight) cats used for experiments with the drug
digitalis had their heart and body weight recorded.  47 of the cats were
female and 97 were male.  The \code{catsM} data frame consists of the data for
the male cats.
}
\section{DATA DESCRIPTION}{
This data frames contain the following columns:


}
\arguments{
\item{Sex}{
A factor for the sex of the cat (levels are \code{F} and \code{M}).
}
\item{Bwt}{
Body weight in kg.
}
\item{Hwt}{
Heart weight in kg.
}}
\section{SOURCE}{
The data were obtained from


Fisher, R.A. (1947) The analysis of covariance method for the relation
between a part and the whole. \emph{Biometrics}, \bold{3}, 65-68.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Venables, W.N. and Ripley, B.D. (1994) 
\emph{Modern Applied Statistics with S-Plus}. Springer-Verlag.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
