\name{catsM}
\alias{catsM}
\title{
Weight Data for Domestic Cats
}
\description{
The \code{catsM} data frame has 97 rows and 3 columns.

144 adult (over 2kg in weight) cats used for experiments with the drug
digitalis had their heart and body weight recorded.  47 of the cats were
female and 97 were male.  The \code{catsM} data frame consists of the data for
the male cats.  The full data are in dataset \code{\link{cats}}
in package \code{MASS}.
}
\usage{
cats
}
\format{
This data frames contain the following columns:

\describe{
\item{\code{Sex}}{
A factor for the sex of the cat (levels are \code{F} and \code{M}).
}
\item{\code{Bwt}}{
Body weight in kg.
}
\item{\code{Hwt}}{
Heart weight in g.
}}}
\seealso{
  \code{\link{cats}}
}
\source{
The data were obtained from

Fisher, R.A. (1947) The analysis of covariance method for the relation
between a part and the whole. \emph{Biometrics}, \bold{3}, 65--68.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Venables, W.N. and Ripley, B.D. (1994) 
\emph{Modern Applied Statistics with S-Plus}. Springer-Verlag.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
