\name{cav}
\alias{cav}
\title{
Position of Muscle Caveolae
}
\description{
The \code{cav} data frame has 138 rows and 2 columns.


The data gives the positions of the individual caveolae in a square region
with sides of length 500 units.  This grid was originally on a 2.65mum
square of muscle fibre.  The data are those points falling in the lower left 
hand quarter of the region used for the
dataset \code{caveolae.dat} in the \pkg{spatial} package by B.D. Ripley (1994).
}
\usage{
cav
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{x}}{
The x coordinate of the caveola's position in the region.
}
\item{\code{y}}{
The y coordinate of the caveola's position in the region.
}}}
\references{
Appleyard, S.T., Witkowski, J.A., Ripley, B.D., Shotton, D.M. and Dubowicz, V.
(1985) A novel procedure for pattern analysis of features present on freeze 
fractured plasma membranes. \emph{Journal of Cell Science}, \bold{74}, 105--117.


Davison, A.C. and Hinkley, D.V. (1997) \emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
