\name{cd4.nested}
\alias{cd4.nested}
\title{
Nested Bootstrap of `cd4' data
}
\section{SUMMARY}{
This is an example of a nested bootstrap for the correlation 
coefficient of the \code{cd4} data frame.  It is used in a practical
in Chapter 5 of Davison and Hinkley (1997).
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\seealso{
\code{\link{cd4}}
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
