\name{cement}
\alias{cement}
\title{
Cement Composition Data
}
\section{SUMMARY}{
The \code{cement} data frame has 13 rows and 5 columns.


13 samples of Portland cement were set. For each sample, the percentages of the
four main chemical ingredients was accurately measured.  While the cement was
setting the amount of heat evolved was also measured.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{x1}{
The percentage by weight of tricalcium aluminate in the sample.
}
\item{x2}{
The percentage by weight of tricalcium silictate in the sample.
}
\item{x3}{
The percentage by weight of tricalcium alumino ferrite in the sample.
}
\item{x4}{
The percentage by weight of dicalcium silictate in the sample.
}
\item{y}{
The heat (in calories per gramme of cement) evolved while the sample was 
setting.
}}
\section{SOURCE}{
The data were obtained from


Woods, H., Steinour, H.H. and Starke, H.R. (1932) Effect of composition of 
Portland cement on heat evolved during hardening. 
\emph{Industrial Engineering and Chemistry}, \bold{24}, 1207-1214.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Venables, W.N. and Ripley, B.D. (1994) 
\emph{Modern Applied Statistics with S-Plus}. Springer-Verlag.
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
