\name{co.transfer}
\alias{co.transfer}
\title{
Carbon Monoxide Transfer
}
\section{SUMMARY}{
The \code{co.transfer} data frame has 7 rows and 2 columns.  Seven smokers with
chickenpox had their levels of carbon monoxide transfer measured on entry
to hospital and then again after 1 week.  The main question being whether
one week of hospitalization  has changed the carbon monoxide transfer factor.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{entry}{
Carbon monoxide transfer factor on entry to hospital.
}
\item{week}{
Carbon monoxide transfer one week after admittance to hospital.
}}
\section{SOURCE}{
The data were obtained from


Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski, E (1994)
\emph{A Handbook of Small Data Sets}. Chapman and Hall.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Ellis, M.E., Neal, K.R. and Webb, A.K. (1987) Is smoking a risk factor for
pneumonia in patients with chickenpox? \emph{British Medical Journal},
\bold{294}, 1002.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
