\name{coal}
\alias{coal}
\title{
Dates of Coal Mining Disasters 
}
\section{SUMMARY}{
The \code{coal} data frame has 191 rows and 1 columns.


This data frame gives the dates of 191 explosions in coal mines which
resulted in 10 or more fatalities.  The time span of the data is from March 15,
1851 until March 22 1962.  
}
\section{DATA DESCRIPTION}{
This data frame contains the following column:


}
\arguments{
\item{date}{
The date of the disaster.  The integer part of \code{date} gives the year.  The day
is represented as the fraction of the year that had elapsed on that day.
}}
\section{SOURCE}{
The data were obtained from


Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski, E. (1994)
\emph{A Handbook of Small Data Sets}, Chapman and Hall.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Jarrett, R.G. (1979) A note on the intervals between coal-mining disasters.
\emph{Biometrika}, \bold{66}, 191-193.
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
