\name{darwin}
\alias{darwin}
\title{
Darwin's Plant Height Differences
}
\section{SUMMARY}{
The \code{darwin} data frame has 15 rows and 1 columns.


Charles Darwin conducted an experiment to examine the superiority of
cross-fertilized plants over self-fertilized plants.  15 pairs of plants
were used. Each pair consisted of one cross-fertilized plant and one
self-fertilized plant which germinated at the same time and grew in the
same pot. The plants were measured at a fixed time after planting and
the difference in heights between the cross- and self-fertilized plants are
recorded in eighths of an inch.
}
\section{DATA DESCRIPTION}{
This data frame contains the following column:


}
\arguments{
\item{y}{
The difference in heights for the pairs of plants (in units of 0.125 inches).
}}
\section{SOURCE}{
The data were obtained from


Fisher, R.A. (1935) \emph{Design of Experiments}. Oliver and Boyd.
}
\section{OTHER REFERENCES}{
Darwin, C. (1876) 
\emph{The Effects of Cross- and Self-fertilisation in the Vegetable Kingdom}. 
John Murray.


Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
