\name{dogs}
\alias{dogs}
\title{ Cardiac Data for Domestic Dogs }
\usage{dogs}
\description{
The \code{dogs} data frame has 7 rows and 2 columns.

Data on the cardiac oxygen consumption and left ventricular
pressure were gathered on 7 domestic dogs.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{mvo}{Cardiac Oxygen Consumption}
    \item{lvp}{Left Ventricular Pressure}
  }
}
\references{
Davison, A. C. and Hinkley, D. V. (1997)
\emph{Bootstrap Methods and Their Application}. Cambridge University
Press.
}
\keyword{datasets}
