\name{fir}
\alias{fir}
\title{
Counts of Balsam-fir Seedlings
}
\section{SUMMARY}{
The \code{fir} data frame has 50 rows and 3 columns.


The number of balsam-fir seedlings in each quadrant of a grid of 50 five foot 
square quadrants were counted.  The grid consisted of 5 rows of 10 quadrants in
each row.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{count}{
The number of seedlings in the quadrant.
}
\item{row}{
The row number of the quadrant. 
}
\item{col}{
The quadrant number within the row.
}}
\section{SOURCE}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
