\name{iden}
\alias{iden}
\alias{zero}
\title{ Construct special matrices}
\description{
  Construct identity or zero matrices of given sizes
}
\usage{
iden(n)
zero(n,m)
}
\arguments{
  \item{n}{Number of rows in the output matrix}
  \item{m}{Number of columns in the output matrix of zero}
}
\value{ A matrix of the appropriate dimensions.}
\author{ Angelo J. Canty}
\keyword{array}