\name{islay}
\alias{islay}
\title{
Jura Quartzite Azimuths on Islay
}
\section{SUMMARY}{
The \code{islay} data frame has 18 rows and 1 columns.


Measurements were taken of paleocurrent azimuths from the Jura Quartzite on
the Scottish island of Islay.
}
\section{DATA DESCRIPTION}{
This data frame contains the following column:


}
\arguments{
\item{theta}{
The angle of the azimuth in degrees East of North.
}}
\section{SOURCE}{
The data were obtained from


Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski, E. (1994)
\emph{A Handbook of Small Data Sets}, Chapman and Hall.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Till, R. (1974) \emph{Statistical Methods for the Earth Scientist}. Macmillan.
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
