\name{lynx}
\alias{lynx}
\title{
Canadian Lynx Trappings
}
\section{SUMMARY}{
This is the same data as contained in the Splus data set of the same name 
except that the time series is now of class \code{"rts"}.  The time series
contains 114 observations.


The data give the annual number of lynx trappings in the Mackenzie River
District of North-West Canada for the period 1821 to 1934.
}
\section{SOURCE}{
The data were obtained from


Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski, E. (1994)
\emph{A Handbook of Small Data Sets}, Chapman and Hall.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) \emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Elton, C. and Nicholson, M. (1942) The ten-year cycle in numbers of lynx in 
Canada. \emph{J. of Animal Ecology}, \bold{11}, 215-244. 
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
