\name{mcycle}
\alias{mcycle}
\title{
Data from a Simulated Motorcycle Accident
}
\description{
The \code{mcycle} data frame has 133 rows and 2 columns.


In an experiment to test crash helmets, a series of measurements of head
acceleration after a simulated motorcycle crash were recorded.
}
\usage{
data(mcycle)
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{times}}{
The time in milliseconds since impact.
}
\item{\code{accel}}{
The recorded head acceleration (in g).
}}}
\source{
The data were obtained from


Silverman, B.W. (1985) Some aspects of the spline smoothing approach to
non-parametric curve fitting. 
\emph{Journal of the Royal Statistical Society, B}, \bold{47}, 1--52.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Venables, W.N. and Ripley, B.D. (1994) 
\emph{Modern Applied Statistics with S-Plus}. Springer-Verlag.
}
\seealso{
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
