\name{neuro}
\alias{neuro}
\title{
Neurophysiological Point Process Data
}
\description{
\code{neuro} is a matrix containing times of observed firing of a neuron in windows
of 250ms either side of the application of a stimulus to a human subject.
Each row of the matrix is a replication of the experiment and there were a 
total of 469 replicates.
}
\note{
There are a lot of missing values in the matrix as different numbers of 
firings 
were observed in different replicates.  The number of firings observed varied 
from 2 to 6.
}
\source{
The data were collected and kindly made available by Dr. S.J. Boniface of the
Neurophysiology Unit at the Radcliffe Infirmary, Oxford.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Ventura, V., Davison, A.C. and Boniface, S.J. (1997) A stochastic model for 
the effect of magnetic brain stimulation on a motorneurone. To appear in
\emph{Applied Statistics}. 
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
