\name{polar}
\alias{polar}
\title{
Pole Positions of New Caledonian Laterites
}
\section{SUMMARY}{
The \code{polar} data frame has 50 rows and 2 columns.


The data are the pole positions from a paleomagnetic study of New Caledonian
laterites.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{lat}{
The latitude (in degrees) of the pole position.  Note that all latitudes
are negative as the axis is taken to be in the lower hemisphere.
}
\item{long}{
The longitude (in degrees) of the pole position.
}}
\section{SOURCE}{
The data were obtained from


Fisher, N.I., Lewis, T. and Embleton, B.J.J. (1987) 
\emph{Statistical Analysis of Spherical Data}. Cambridge University Press.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
