\name{remission}
\alias{remission}
\title{
Cancer Remission and Cell Activity
}
\section{SUMMARY}{
The \code{remission} data frame has 27 rows and 3 columns.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{LI}{
A measure of cell activity.
}
\item{m}{
The number of patients in each group (all values are actually 1 here).
}
\item{r}{
The number of patients (out of \code{m}) who went into remission.
}}
\section{SOURCE}{
The data were obtained from


Freeman, D.H. (1987) \emph{Applied Categorical Data Analysis}. Marcel Dekker.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
