\name{survival}
\alias{survival}
\title{
Survival of Rats after Radiation Doses
}
\section{SUMMARY}{
The \code{survival} data frame has 14 rows and 2 columns.


The data measured the survival percentages of batches of rats who were given 
varying doses of radiation.  At each of 6 doses there were two or three
replications of the experiment.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{dose}{
The dose of radiation administered (rads).
}
\item{surv}{
The survival rate of the batches expressed as a percentage.
}}
\section{SOURCE}{
The data were obtained from


Efron, B. (1988) Computer-intensive methods in statistical regression.
\emph{SIAM Review}, \bold{30}, 421-449.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
