\name{tsboot-internal}
\alias{make.ends}
\alias{scramble}
\alias{ts.array}
\title{Time Series Bootstrap Functions}
\description{
  Internal functions for use with \code{tsboot}.}
\usage{
make.ends(a, n)
scramble(ts, norm = TRUE)
}
\details{
  These functions are internal functions used by \code{tsboot}.
They should not be called directly by the user.}
\author{ Angelo J. Canty}
\seealso{\code{\link{tsboot}}}
\keyword{internal} 
