\name{tuna}
\alias{tuna}
\title{
Tuna Sighting Data
}
\section{SUMMARY}{
The \code{tuna} data frame has 64 rows and 1 columns.  


The data come from an aerial
line transect survey of Southern Bluefin Tuna in the Great Australian Bight.
An aircraft with two spotters on board flies randomly allocated line transects.
Each school of tuna sighted is counted and its perpendicular distance from the
transect measured.  The survey was conducted in summer when tuna tend to stay
on the surface.
}
\section{DATA DESCRIPTION}{
This data frame contains the following column:


}
\arguments{
\item{y}{
The perpendicular distance, in miles, from the transect for 64 independent 
sightings of tuna schools.
}}
\section{SOURCE}{
The data were obtained from


Chen, S.X. (1996) Empirical likelihood confidence intervals for
nonparametric density estimation. \emph{Biometrika}, \bold{83}, 329-341.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
