\name{urine}
\alias{urine}
\title{
Urine Analysis Data
}
\section{SUMMARY}{
The \code{urine} data frame has 79 rows and 7 columns.


79 urine specimens were analyzed in an effort to determine if certain physical
characteristics of the urine might be related to the formation of calcium
oxalate crystals.  
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{r}{
Indicator of the presence of calcium oxalate crystals.
}
\item{gravity}{
The specific gravity of the urine.
}
\item{ph}{
The pH reading of the urine.
}
\item{osmo}{
The osmolarity of the urine. Osmolarity is proportional to the concentration
of molecules in solution.
}
\item{cond}{
The conductivity of the urine.  Conductivity is proportional to the 
concentration of charged ions in solution.
}
\item{urea}{
The urea concentration in millimoles per litre.
}
\item{calc}{
The calcium concentration in millimoles per litre.
}}
\section{SOURCE}{
The data were obtained from


Andrews, D.F. and Herzberg, A.M. (1985) 
\emph{Data: A Collection of Problems from Many Fields for the Student and Research Worker}. Springer-Verlag.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
