\name{wool}
\alias{wool}
\title{
Australian Relative Wool Prices
}
\section{SUMMARY}{
\code{wool} is a time series of class \code{"rts"} and contains 309 observations.


Each week that the market is open the Australian Wool Corporation set a floor
price which determines their policy on intervention and is therefore a 
reflection of the overall price of wool for the week in question.  Actual prices
paid can vary considerably about the floor price.  The series here is the log
of the ratio between the price for fine grade wool and the floor price, each
market week between July 1976 and Junk 1984.
}
\section{SOURCE}{
The data were obtained from


Diggle, P.J. (1990) \emph{Time Series: A Biostatistical Introduction}.
Oxford University Press.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) \emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{sysdata}
% Converted by Sd2Rd version 0.3-1.
