\name{brambles}
\alias{brambles}
\title{
Spatial Location of Bramble Canes
}
\description{
The \code{brambles} data frame has 823 rows and 3 columns.


The location of living bramble canes in a 9m square plot was recorded.  
We take 9m to be the unit of distance so that the plot can be thought of
as a unit square. The bramble canes were also classified by their age.
}
\usage{
brambles
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{x}}{
The x coordinate of the position of the cane in the plot.
}
\item{\code{y}}{
The y coordinate of the position of the cane in the plot.
}
\item{\code{age}}{
The age classification of the canes; \code{0} indicates a newly emerged cane,
\code{1} indicates a one year old cane and \code{2} indicates a two year old cane.
}}}
\source{
The data were obtained from

Diggle, P.J. (1983) \emph{Statistical Analysis of Spatial Point Patterns}.
Academic Press.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
