\name{capability}
\alias{capability}
\title{
Simulated Manufacturing Process Data
}
\description{
The \code{capability} data frame has 75 rows and 1 columns.


The data are simulated successive observations from a process in equilibrium.
The process is assumed to have specification limits (5.49, 5.79).
}
\usage{
capability
}
\format{
This data frame contains the following column:


\describe{
\item{\code{y}}{
The simulated measurements.
}}}
\source{
The data were obtained from


Bissell, A.F. (1990) How reliable is your capability index?
\emph{Applied Statistics}, \bold{39}, 331--340.
}
\references{
Canty, A.J. and Davison, A.C. (1996) Implementation of saddlepoint 
approximations to resampling distributions. To appear in 
\emph{Computing Science and Statistics; Proceedings of the 28th Symposium on the Interface}.


Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
