\name{cloth}
\alias{cloth}
\title{
Number of Flaws in Cloth
}
\description{
The \code{cloth} data frame has 32 rows and 2 columns.
}
\usage{
cloth
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{x}}{
The length of the roll of cloth.
}
\item{\code{y}}{
The number of flaws found in the roll.
}}}
\source{
The data were obtained from


Bissell, A.F. (1972) A negative binomial model with varying element size.
\emph{Biometrika}, \bold{59}, 435--441.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
