\name{fir}
\alias{fir}
\title{
Counts of Balsam-fir Seedlings
}
\description{
The \code{fir} data frame has 50 rows and 3 columns.


The number of balsam-fir seedlings in each quadrant of a grid of 50 five foot 
square quadrants were counted.  The grid consisted of 5 rows of 10 quadrants in
each row.
}
\usage{
fir
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{count}}{
The number of seedlings in the quadrant.
}
\item{\code{row}}{
The row number of the quadrant. 
}
\item{\code{col}}{
The quadrant number within the row.
}}}
\source{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
