\name{frets}
\alias{frets}
\title{
Head Dimensions in Brothers
}
\description{
The \code{frets} data frame has 25 rows and 4 columns.


The data consist of measurements of the length and breadth of the heads of 
pairs of adult brothers in 25 randomly sampled families.  All measurements
are expressed in millimetres.
}
\usage{
frets
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{l1}}{
The head length of the eldest son.
}
\item{\code{b1}}{
The head breadth of the eldest son.
}
\item{\code{l2}}{
The head length of the second son.
}
\item{\code{b2}}{
The head breadth of the second son.
}}}
\source{
The data were obtained from


Frets, G.P. (1921) Heredity of head form in man. \emph{Genetica}, \bold{3},
193.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Whittaker, J. (1990) \emph{Graphical Models in Applied Multivariate Statistics}.
John Wiley.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
