\name{islay}
\alias{islay}
\title{
Jura Quartzite Azimuths on Islay
}
\description{
The \code{islay} data frame has 18 rows and 1 columns.


Measurements were taken of paleocurrent azimuths from the Jura Quartzite on
the Scottish island of Islay.
}
\usage{
islay
}
\format{
This data frame contains the following column:


\describe{
\item{\code{theta}}{
The angle of the azimuth in degrees East of North.
}}}
\source{
The data were obtained from


Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski, E. (1994)
\emph{A Handbook of Small Data Sets}, Chapman and Hall.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Till, R. (1974) \emph{Statistical Methods for the Earth Scientist}. Macmillan.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
