\name{mammals}
\alias{mammals}
\title{
Brain and Body Weights in Mammals
}
\description{
The \code{mammals} data frame has 62 rows and 2 columns.


The data give average brain weight and average body weight for 62 different species of land mammals.  Each row in the data frame is named with the name
of the species whose measurements are in the row.
}
\usage{
data(mammals)
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{body}}{
The average body weight for the species (in kilogrammes).
}
\item{\code{brain}}{
The average brain weight for the species (in grammes).
}}}
\source{
The data were obtained from


Weisberg, S. (1985) \emph{Applied Linear Regression} (Second edition).
John Wiley.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Venables, W.N. and Ripley, B.D. (1994) 
\emph{Modern Applied Statistics with S-Plus}. Springer-Verlag.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
