\name{mcycle}
\alias{mcycle}
\title{
Data from a Simulated Motorcycle Accident
}
\section{SUMMARY}{
The \code{mcycle} data frame has 133 rows and 2 columns.


In an experiment to test crash helmets, a series of measurements of head
acceleration after a simulated motorcycle crash were recorded.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:
}
\arguments{
\item{times}{
The time in milliseconds since impact.
}
\item{accel}{
The recorded head acceleration (in g).
}}
\section{SOURCE}{
The data were obtained from


Silverman, B.W. (1985) Some aspects of the spline smoothing approach to
non-parametric curve fitting. 
\emph{Journal of the Royal Statistical Society, B}, \bold{47}, 1-52.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Venables, W.N. and Ripley, B.D. (1994) 
\emph{Modern Applied Statistics with S-Plus}. Springer-Verlag.
}
\seealso{
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
