\name{phones}
\alias{phones}
\title{
Belgium Phone Calls 1950-1973
}
\section{SUMMARY}{
The \code{phones} data frame has 24 rows and 2 columns.
}
\section{DATA DESCRIPTION}{
This data frame contains the following columns:


}
\arguments{
\item{year}{
The last two digits of the year.
}
\item{calls}{
The number of telephone calls made (in millions of calls).
}}
\section{SOURCE}{
The data were obtained from


Rousseeuw, P.J. and Leroy, A.M. (1987) 
\emph{Robust Regression and Outlier Detection}. John Wiley.
}
\section{OTHER REFERENCES}{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.


Venables, W.N. and Ripley, B.D. (1994) 
\emph{Modern Applied Statistics with Splus}. Springer-Verlag
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-1.
