\name{polar}
\alias{polar}
\title{
Pole Positions of New Caledonian Laterites
}
\description{
The \code{polar} data frame has 50 rows and 2 columns.


The data are the pole positions from a paleomagnetic study of New Caledonian
laterites.
}
\usage{
polar
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{lat}}{
The latitude (in degrees) of the pole position.  Note that all latitudes
are negative as the axis is taken to be in the lower hemisphere.
}
\item{\code{long}}{
The longitude (in degrees) of the pole position.
}}}
\source{
The data were obtained from


Fisher, N.I., Lewis, T. and Embleton, B.J.J. (1987) 
\emph{Statistical Analysis of Spherical Data}. Cambridge University Press.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
