\name{remission}
\alias{remission}
\title{
Cancer Remission and Cell Activity
}
\description{
The \code{remission} data frame has 27 rows and 3 columns.
}
\usage{
remission
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{LI}}{
A measure of cell activity.
}
\item{\code{m}}{
The number of patients in each group (all values are actually 1 here).
}
\item{\code{r}}{
The number of patients (out of \code{m}) who went into remission.
}}}
\source{
The data were obtained from


Freeman, D.H. (1987) \emph{Applied Categorical Data Analysis}. Marcel Dekker.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
