\name{urine}
\alias{urine}
\title{
Urine Analysis Data
}
\description{
The \code{urine} data frame has 79 rows and 7 columns.


79 urine specimens were analyzed in an effort to determine if certain physical
characteristics of the urine might be related to the formation of calcium
oxalate crystals.  
}
\usage{
urine
}
\format{
This data frame contains the following columns:


\describe{
\item{\code{r}}{
Indicator of the presence of calcium oxalate crystals.
}
\item{\code{gravity}}{
The specific gravity of the urine.
}
\item{\code{ph}}{
The pH reading of the urine.
}
\item{\code{osmo}}{
The osmolarity of the urine. Osmolarity is proportional to the concentration
of molecules in solution.
}
\item{\code{cond}}{
The conductivity of the urine.  Conductivity is proportional to the 
concentration of charged ions in solution.
}
\item{\code{urea}}{
The urea concentration in millimoles per litre.
}
\item{\code{calc}}{
The calcium concentration in millimoles per litre.
}}}
\source{
The data were obtained from


Andrews, D.F. and Herzberg, A.M. (1985) 
\emph{Data: A Collection of Problems from Many Fields for the Student and Research Worker}. Springer-Verlag.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) 
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.15.
